requirements.txt
azure-storage-blob==12.16.0

